requirements.txt
surya-ocr
transformers
pdf2image
Pillow